+++ instrumentation/opentelemetry-instrumentation-pymongo/src/opentelemetry/instrumentation/pymongo/__init__.py
@@ -121,10 +121,10 @@ class CommandTracer(monitoring.CommandListener):
         ):
             return
         command = event.command.get(event.command_name, "")
-        name = event.command_name
+        name = event.database_name
+        name += "." + event.command_name
         statement = event.command_name
         if command:
-            name += "." + str(command)
             statement += " " + str(command)
 
         try: